#43: Datamodel: introduce voter groups and move RS256 keys in this table
Introduced voter groups for the case that an election creator has misplaced / deleted JWTs and wants to invalidate them. As they aren't able to look into the payload of the tokens in the described case this needs to be possible in the UI. For this voter groups need to be stored inside the database.

Moved RS256 keys into voterGroups table as previously it would have been unclear which key was used for a voters JWT. A voter may vote in multiple elections of which each previously had their own RS256 keys.

Now if a voter provides their JWT, it is automatically clear which key should be the one used to generate it. It also is possible to invalidate only a specific group of voters and not all voters associated with one election.

Signed-off-by: David Sievers <david.sievers@uni-ulm.de>