#271: change the bigints in encryptedFilledBallotPaper to strings
This is necessary as JSON serialization does not work with bigint.
Effectively this would mean for sending and persisting the data, bigint needs to be converted to a string.
When one wants to work with them, they need to be turned back.
The way it is now, working with the bigint values still requires a cast but the sending and persisting requires no extra action.

Signed-off-by: GitHub <noreply@github.com>